:white_check_mark: Remove old tests
