+++ requirements/base.in
@@ -19,4 +19,4 @@ djangorestframework-camel-case
 djangorestframework-filters
 drf-yasg
 drf-writable-nested
-vng-api-common[notifications]==0.42.0
+vng-api-common[notifications]==0.42.1
