+++ requirements/base.txt
@@ -50,4 +50,4 @@ six==1.11.0               # via django-markup, djangorestframework-gis, drf-yasg
 unidecode==1.0.22         # via vng-api-common
 uritemplate==3.0.0        # via coreapi, drf-yasg
 urllib3==1.24.3           # via requests
-vng-api-common[notifications]==0.42.0
+vng-api-common[notifications]==0.42.1
