+++ requirements/jenkins.txt
@@ -68,7 +68,7 @@ text-unidecode==1.2       # via faker
 unidecode==1.0.22
 uritemplate==3.0.0
 urllib3==1.24.3
-vng-api-common[notifications]==0.42.0
+vng-api-common[notifications]==0.42.1
 waitress==1.1.0           # via webtest
 webob==1.8.2              # via webtest
 webtest==2.0.30
