+++ src/zrc/api/tests/test_auth.py
@@ -1,92 +1,136 @@
 """
 Guarantee that the proper authorization amchinery is in place.
 """
-from unittest import skip
+from django.test import override_settings
 
 from rest_framework import status
 from rest_framework.test import APITestCase
-from vng_api_common.tests import AuthCheckMixin, generate_jwt
+from vng_api_common.constants import VertrouwelijkheidsAanduiding
+from vng_api_common.tests import AuthCheckMixin, JWTAuthMixin
 
-from zrc.datamodel.tests.factories import ZaakFactory
+from zrc.datamodel.tests.factories import (
+    StatusFactory, ZaakFactory, ZaakObjectFactory
+)
+from zrc.tests.utils import ZAAK_READ_KWARGS
 
 from ..scopes import SCOPE_ZAKEN_ALLES_LEZEN
 from .utils import reverse
 
 
-class ZakenCreateTests(AuthCheckMixin, APITestCase):
+@override_settings(ZDS_CLIENT_CLASS='vng_api_common.mocks.MockClient')
+class ZakenScopeForbiddenTests(AuthCheckMixin, APITestCase):
 
     def test_cannot_create_zaak_without_correct_scope(self):
         url = reverse('zaak-list')
-
         self.assertForbidden(url, method='post')
-
-
-@skip('Test old Authentication Format. New tests in test_auth_new_format.py')
-class ZakenReadTests(AuthCheckMixin, APITestCase):
 
     def test_cannot_read_without_correct_scope(self):
         zaak = ZaakFactory.create()
+        status = StatusFactory.create()
+        zaak_object = ZaakObjectFactory.create()
         urls = [
             reverse('zaak-list'),
             reverse('zaak-detail', kwargs={'uuid': zaak.uuid}),
             reverse('status-list'),
-            reverse('status-detail', kwargs={'uuid': 'dummy'}),
+            reverse('status-detail', kwargs={'uuid': status.uuid}),
             reverse('zaakobject-list'),
-            reverse('zaakobject-detail', kwargs={'uuid': 'dummy'}),
+            reverse('zaakobject-detail', kwargs={'uuid': zaak_object.uuid}),
         ]
 
         for url in urls:
             with self.subTest(url=url):
-                self.assertForbidden(url, method='get')
+                self.assertForbidden(url, method='get', request_kwargs=ZAAK_READ_KWARGS)
 
-    def test_zaaktypes_claim(self):
+
+class ZaakReadCorrectScopeTests(JWTAuthMixin, APITestCase):
+    scopes = [SCOPE_ZAKEN_ALLES_LEZEN]
+    zaaktype = 'https://zaaktype.nl/ok'
+    max_vertrouwelijkheidaanduiding = VertrouwelijkheidsAanduiding.openbaar
+
+    def test_zaak_list(self):
         """
-        Assert you can only read ZAAKen of the zaaktypes in the claim.
+        Assert you can only list ZAAKen of the zaaktypes and vertrouwelijkheidaanduiding
+        of your authorization
         """
-        ZaakFactory.create(zaaktype='https://zaaktype.nl/ok')
-        ZaakFactory.create(zaaktype='https://zaaktype.nl/not_ok')
-        url = reverse('zaak-list')
-        jwt = generate_jwt(
-            scopes=[SCOPE_ZAKEN_ALLES_LEZEN],
-            zaaktypes=['https://zaaktype.nl/ok'],
+        ZaakFactory.create(
+            zaaktype='https://zaaktype.nl/ok',
+            vertrouwelijkheidaanduiding=VertrouwelijkheidsAanduiding.openbaar
         )
-        self.client.credentials(HTTP_AUTHORIZATION=jwt)
+        ZaakFactory.create(
+            zaaktype='https://zaaktype.nl/not_ok',
+            vertrouwelijkheidaanduiding=VertrouwelijkheidsAanduiding.openbaar
+        )
+        ZaakFactory.create(
+            zaaktype='https://zaaktype.nl/ok',
+            vertrouwelijkheidaanduiding=VertrouwelijkheidsAanduiding.zeer_geheim
+        )
+        ZaakFactory.create(
+            zaaktype='https://zaaktype.nl/not_ok',
+            vertrouwelijkheidaanduiding=VertrouwelijkheidsAanduiding.zeer_geheim
+        )
+        url = reverse('zaak-list')
 
         response = self.client.get(url, HTTP_ACCEPT_CRS='EPSG:4326')
 
         self.assertEqual(response.status_code, status.HTTP_200_OK)
+
         results = response.data['results']
+
         self.assertEqual(len(results), 1)
         self.assertEqual(results[0]['zaaktype'], 'https://zaaktype.nl/ok')
+        self.assertEqual(results[0]['vertrouwelijkheidaanduiding'], VertrouwelijkheidsAanduiding.openbaar)
 
-    def test_zaaktypes_claim_detail(self):
+    def test_zaak_retreive(self):
         """
-        Assert you can only read ZAAKen of the zaaktypes in the claim.
+        Assert you can only read ZAAKen of the zaaktypes and vertrouwelijkheidaanduiding
+        of your authorization
         """
-        zaak = ZaakFactory.create(zaaktype='https://zaaktype.nl/not_ok')
-        url = reverse('zaak-detail', kwargs={'uuid': zaak.uuid})
-
-        self.assertForbiddenWithCorrectScope(
-            url, [SCOPE_ZAKEN_ALLES_LEZEN],
-            zaaktypes=['https://zaaktype.nl/ok'],
-            request_kwargs={'HTTP_ACCEPT_CRS': 'EPSG:4326'}
+        zaak1 = ZaakFactory.create(
+            zaaktype='https://zaaktype.nl/not_ok',
+            vertrouwelijkheidaanduiding=VertrouwelijkheidsAanduiding.openbaar
         )
-
-    def test_zaaktypes_wildcard(self):
-        zaak = ZaakFactory.create()
-
-        list_url = reverse('zaak-list')
-        detail_url = reverse('zaak-detail', kwargs={'uuid': zaak.uuid})
-
-        jwt = generate_jwt(
-            scopes=[SCOPE_ZAKEN_ALLES_LEZEN],
-            zaaktypes=['*'],
+        zaak2 = ZaakFactory.create(
+            zaaktype='https://zaaktype.nl/ok',
+            vertrouwelijkheidaanduiding=VertrouwelijkheidsAanduiding.zeer_geheim
         )
-        self.client.credentials(HTTP_AUTHORIZATION=jwt)
+        url1 = reverse('zaak-detail', kwargs={'uuid': zaak1.uuid})
+        url2 = reverse('zaak-detail', kwargs={'uuid': zaak2.uuid})
 
-        list_response = self.client.get(list_url, HTTP_ACCEPT_CRS='EPSG:4326')
-        detail_response = self.client.get(detail_url, HTTP_ACCEPT_CRS='EPSG:4326')
+        response1 = self.client.get(url1, **ZAAK_READ_KWARGS)
+        response2 = self.client.get(url2, **ZAAK_READ_KWARGS)
 
-        self.assertEqual(list_response.status_code, status.HTTP_200_OK)
-        self.assertEqual(detail_response.status_code, status.HTTP_200_OK)
-        self.assertEqual(len(list_response.data['results']), 1)
+        self.assertEqual(response1.status_code, status.HTTP_403_FORBIDDEN)
+        self.assertEqual(response2.status_code, status.HTTP_403_FORBIDDEN)
+
+    def test_read_superuser(self):
+        """
+        superuser read everything
+        """
+        self.applicatie.heeft_alle_autorisaties = True
+        self.applicatie.save()
+
+        ZaakFactory.create(
+            zaaktype='https://zaaktype.nl/ok',
+            vertrouwelijkheidaanduiding=VertrouwelijkheidsAanduiding.openbaar
+        )
+        ZaakFactory.create(
+            zaaktype='https://zaaktype.nl/not_ok',
+            vertrouwelijkheidaanduiding=VertrouwelijkheidsAanduiding.openbaar
+        )
+        ZaakFactory.create(
+            zaaktype='https://zaaktype.nl/ok',
+            vertrouwelijkheidaanduiding=VertrouwelijkheidsAanduiding.zeer_geheim
+        )
+        ZaakFactory.create(
+            zaaktype='https://zaaktype.nl/not_ok',
+            vertrouwelijkheidaanduiding=VertrouwelijkheidsAanduiding.zeer_geheim
+        )
+        url = reverse('zaak-list')
+
+        response = self.client.get(url, HTTP_ACCEPT_CRS='EPSG:4326')
+
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
+
+        results = response.data['results']
+
+        self.assertEqual(len(results), 4)
